Improve rule rendering, nested lists
This solves a problem in which nested lists were not being
appropriately expanded in the context of rules.

Previously, given:

(def rule [:h1 '({:foo "bar"})])

calling `(render-css rule)` would produce `""`. This is now correctly
rendered as `"h1{foo:bar}"` thanks to an improved rule processing
function `divide-rule`.

`divid-rule` works by splitting the rule in to a triple consisting of
the selector, it's declarations, and subrules. The divided result can
then be easily consumed by `render-rule` and eliminates the original
task of both filtering maps (declarations) and then vectors (subrules).