Add second argument to defunit, percentage unit
This patch allows a second agument to be passed to `defunit` as the CSS
unit when rendereding occurs. It allows one to define a set of unit
functions with similar prefix but use a different unit "under the hood".
With it, it is possible to now have a builtin percentage function:

  (defunit percentage :%)

  (css [:body {:font-size (percentage 100)}])
  ;; => body{font-size:100%}

This is especially handy in this case since identifiers including a
`%` character can be problematic with the Clojure reader.