Use `seq?` instead of `list?`
This patch fixes a subtle problem in which types such as
clojure.lang.ArraySeq were not being rendered. Consider the following
stylesheet code:

  (defn on-hover [& styles]
    ["&:hover" styles])

  (css [:a (on-hover {:font-weight "bold"})])

The expected output should be:

  "a:hover{font-weight:bold}"

but it is actually "". Prior to this patch the `on-hover` function
would have to be written as:

  (defn on-hover [& styles]
    (into ["&:hover"] styles))

which is tedious and can lead very dirty code.