Initial work on generic arithemetic
Garden needs generic arithmetic because it can be a pain to build
complex stylesheets without it. Often when working with functions that
accept units or numbers, tedious type checking and manual conversion is
involved. (I've found this to be the case implementing something as
simple as a 960 grid system.)

These functions will remove such nonsense making it easier for
stylesheet authors to build more elaborate functions without the hassle.

This commit adds the + and - operators. * and / to come.