Replace CSS{Import,Keyframes,MediaQuery} with CSSAtRule
Since CSS has several at-rules, and it would be nice if Garden could
target all of them, it makes sense to have a single type which
represents them.

The constructor requires an identifier and value. The identifier being
the name of the at-rule and the value being any type (although this is
usually will be a map).

The expansion and rendering process can now simply dispatch on the
CSSAtRule type which in turn uses a multimethod to determine the
correct action.

Perhaps at some point the same sort of refactoring could be applied to
the remaining Garden types (CSSUnit and CSSFunction).