Redesign
This patch rethinks the architecture of Garden in terms of well
defined boundaries between parsing and compiling. Instead of
directly compiling objects into CSS, Garden will first parse the objects
into a CSS AST which can then be consumed by the compiler. This
architecture is closer to what one might expect to find in a
programming language implementation setting. This new architecture
yields greater flexibility to introduce new features and semantics
while vastly simplifying the work of the compiler.

As a result many requested features and solutions to bugs have now
been either implemented, solved, or are now trivial to implement or
solve.

Changelog updates forthcoming.

Closes #31
Closes #109
Closes #115
Closes #116
Closes #119
Closes #120
Closes #121