Add support for Map-like collections inside styles
By changing `{}` with `(empty d)`, we gain the ability to use different Map-like
collections in the style definition. This is relevant because ordering can be
important within the css block itself, and clojure maps are not ordered by
definition.

The tests take into account that Clojure's threshold to choose between
array-map (naturally ordered) and hash-map (unordered) is 8 elements, and also
that an explicit array-map can take one more element.

The tests also verify that the underlying collection properties are preserved
during processing, such as the sorting comparator. Although not tested, metadata
should also be preserved.