Changes clamp -> util/clip
Previous commits did not recognize the existing util/clip function. This
removes the unnecessary clamp function that was introduced.

Additionally, scale-color-value was updated to accept min-val and
max-val arguments to inform the util/clip function in the case that
values should be clamped to something other than [0, 100]. This could
happen in a function like scale-alpha that clamped from [0.0, 1.0].